Refactor primary id to empId as Integer

+++ EmployeeAddress/src/main/java/employeeaddress/service/EmployeeAddressRepository.java
@@ -4,6 +4,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import employeeaddress.model.EmployeeAddress;
 
-public interface EmployeeAddressRepository extends JpaRepository<EmployeeAddress, Long> {
+public interface EmployeeAddressRepository extends JpaRepository<EmployeeAddress, Integer> {
 
 }
