+++ EmployeeDetail/src/main/java/employeedetail/service/EmployeeDetailRepository.java
@@ -4,6 +4,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import employeedetail.model.EmployeeDetail;
 
-public interface EmployeeDetailRepository extends JpaRepository<EmployeeDetail, Long> {
+public interface EmployeeDetailRepository extends JpaRepository<EmployeeDetail, Integer> {
 
 }
