+++ EmployeeProject/src/main/java/employeeproject/service/EmployeeProjectRepository.java
@@ -4,6 +4,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import employeeproject.model.EmployeeProject;
 
-public interface EmployeeProjectRepository extends JpaRepository<EmployeeProject, Long> {
+public interface EmployeeProjectRepository extends JpaRepository<EmployeeProject, Integer> {
 
 }
